Introduced geometric base clases for lines, circles and arcs
